+++ apps/fluent-tester/src/FluentTester/TestComponents/Avatar/E2EAvatarTest.tsx
@@ -1,10 +1,10 @@
-import { JSAvatar } from '@fluentui-react-native/experimental-avatar';
+import { Avatar } from '@fluentui-react-native/avatar';
 import * as React from 'react';
 import { View } from 'react-native';
 import {
-  JSAVATAR_ACCESSIBILITY_LABEL,
-  JSAVATAR_ACCESSIBILITY_HINT,
-  JSAVATAR_ACCESSIBILITY_ROLE,
+  AVATAR_ACCESSIBILITY_LABEL,
+  AVATAR_ACCESSIBILITY_HINT,
+  AVATAR_ACCESSIBILITY_ROLE,
   AVATAR_TEST_COMPONENT,
   AVATAR_SECONDARY_TEST_COMPONENT,
 } from './consts';
@@ -12,13 +12,13 @@ import {
 export const E2EAvatarTest: React.FunctionComponent = () => {
   return (
     <View>
-      <JSAvatar
+      <Avatar
         testID={AVATAR_TEST_COMPONENT}
-        accessibilityHint={JSAVATAR_ACCESSIBILITY_HINT}
-        accessibilityLabel={JSAVATAR_ACCESSIBILITY_LABEL}
-        accessibilityRole={JSAVATAR_ACCESSIBILITY_ROLE}
+        accessibilityHint={AVATAR_ACCESSIBILITY_HINT}
+        accessibilityLabel={AVATAR_ACCESSIBILITY_LABEL}
+        accessibilityRole={AVATAR_ACCESSIBILITY_ROLE}
       />
-      <JSAvatar testID={AVATAR_SECONDARY_TEST_COMPONENT} name="Richard" badge={{ status: 'available' }} />
+      <Avatar testID={AVATAR_SECONDARY_TEST_COMPONENT} name="Richard" badge={{ status: 'available' }} />
     </View>
   );
 };